requirements.txt
fastapi
requests
openai
pydantic
uvicorn
tenacity
aiohttp
pydantic-settings
google-genai
pandas